pipeline/requirements.txt
accelerate==1.12.0
aistudio-sdk==0.3.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==16.0.1
bce-python-sdk==0.9.53
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chandra-ocr==0.1.8
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.3
distro==1.9.0
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdownify==1.1.0
MarkupSafe==3.0.3
modelscope==1.32.0
mpmath==1.3.0
narwhals==2.12.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.8.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddlepaddle==3.2.0
paddlex==3.3.9
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.29.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1

